TransitionsInTimeseries.jl

TransitionsInTimeseries — Module

TransitionsInTimeseries.jl

A Julia package for estimating transitions (from one dynamic regime or stable state to another) in timeseries and testing the statistical significance of found transitions. It integrates with the entire Julia ecosystem of timeseries analysis, and hence offers thousands of metrics that can indicate transitions right out of the box. It also offers a variety of analysis pipelines for identifying transitions and a variety of statistical pipelines for testing for significance.

In contrast to other existing software with similar target application, TransitionsInTimeseries.jl defines a generic interface for how to find transitions and how to test for significance. Within this interface, it is easy to expand the software in three orthogonal ways:

Add new indicators that work with the existing analysis pipelines for finding transitions

Add new analysis pipelines for finding transitions

Add new ways for testing whether already found transitions are significant

This package is currently under active development and not yet registered in the Julia general registry. To install it, first go into package-manager mode in the Julia REPL (press ]) and then run

add https://github.com/JuliaDynamics/TransitionsInTimeseries.jl

All further information is provided in the documentation, which you can either find online or build locally by running the docs/make.jl file.

Alternative names for this package could have been: Early Warning Signals / Resilience Indicators / Regime-Shift Identifiers / Change-Point Detectors, or however else you want to call them!

Star us on GitHub!

If you have found this package useful, please consider starring it on GitHub. This gives us an accurate lower bound of the (satisfied) user count.

Content

Multi-stable systems can display abrupt transitions between two stability regimes. To predict such transitions in real-world systems solely based on data, mathematical tools have been developed in the last decades. Numerous terminologies have been used for them, such as early warning signals, resilience indicators, regime-shift identifiers, change-point detection and transition indicators. TransitionsInTimeseries.jl sticks to the latter terminology and provides an interface that:

Allows a fast computation of common transition indicators with a couple of lines, as demonstrated in the example section.

Makes the surrogate analysis to test for significance under the hub.

Can be easily extended by any user without touching the source code.

Reduces the programming overhead for any researcher willing to benchmark new methods.

Eases the reproducibility thanks to a clear syntax, a simple installation and RNG-seeded surrogate generation.

Increases trustworthiness thanks to a large test suite.

Similar projects

An R toolbox and a Python library already exist. However, we believe that they are difficult to extend for the user. Furthermore, they do not offer a native performant code, as here allowed by the use of Julia.

Approaches

Over the last decades, research on transition indicators has largely focused on Critical Slowing Down (CSD). CSD is observed when a system approaches a Hopf, a transcritical or a fold bifurcation and consists in a resilience loss of the system. For instance this can be diagnosed by an increase of the variance and the AR1-regression coefficient, as demonstrated in the example section. However, we emphasize that this is one out of many possible approaches for obtaining transition indicators. Recent work has explored new approaches relying on nonlinear dynamics or machine learning. TransitionsInTimeseries.jl is designed to allow these cutting-edge methods and foster the development of new ones.