PeriodicOrbits.jl

PeriodicOrbits — Module

PeriodicOrbits.jl

PeriodicOrbits.jl provides both the interface and algorithm implementations for finding stable and unstable periodic orbits of discrete and continuous time dynamical systems based on the DynamicalSystems.jl ecosystem.

It can be used as a standalone package, or as part of DynamicalSystems.jl.

To install it, run import Pkg; Pkg.add("PeriodicOrbits").

In the future this package will be a basis for local continuation integrated with DynamicalSystems.jl.

All further information is provided in the documentation, which you can either find online or build locally by running the docs/make.jl file.

Note: Throughout the documentation, "Periodic Orbit" is often abbreviated as "PO" for brevity.