Utility systems

TimeseriesSurrogates.SNLST — Function

SNLST(n_steps, x₀, k)

Dynamically linear process transformed by a strongly nonlinear static transformation (SNLST)[1].

Equations

The system is by the following map:

\[x(t) = k x(t-1) + a(t)\]

with the transformation $s(t) = x(t)^3$.

References

TimeseriesSurrogates.randomwalk — Function

randomwalk(n_steps, x₀)

Linear random walk (AR(1) process with a unit root)[1]. This is an example of a nonstationary linear process.

References

TimeseriesSurrogates.NSAR2 — Function

NSAR2(n_steps, x₀, x₁)

Cyclostationary AR(2) process[1].

References

TimeseriesSurrogates.AR1 — Function

AR1(; n_steps, x₀, k, rng)

Simple AR(1) model given by the following map:

\[x(t+1) = k x(t) + a(t),\]

where $a(t)$ is a draw from a normal distribution with zero mean and unit variance. x₀ sets the initial condition and k is the tunable parameter in the map. rng is a random number generator

TimeseriesSurrogates.random_cycles — Function

random_cycles(; periods=10 dt=π/20, σ = 0.05, frange = (0.8, 2.0))

Make a timeseries that is composed of period full sine wave periods, each with a random frequency in the range given by frange, and added noise with std σ. The sampling time is dt.

1Lucio et al., Phys. Rev. E 85, 056202 (2012). https://journals.aps.org/pre/abstract/10.1103/PhysRevE.85.056202

1Lucio et al., Phys. Rev. E 85, 056202 (2012). https://journals.aps.org/pre/abstract/10.1103/PhysRevE.85.056202

1Lucio et al., Phys. Rev. E 85, 056202 (2012). https://journals.aps.org/pre/abstract/10.1103/PhysRevE.85.056202