Probabilities

Entropies.probabilities — Method

Permutation-based symbol probabilities

probabilities(x::Dataset, est::SymbolicPermutation) → Vector{<:Real} 
probabilities(x::AbstractVector, est::SymbolicPermutation;  m::Int = 2, τ::Int = 1) → Vector{<:Real} 

probabilities!(s::Vector{Int}, x::Dataset, est::SymbolicPermutation) → Vector{<:Real} 
probabilities!(s::Vector{Int}, x::AbstractVector, est::SymbolicPermutation;  m::Int = 2, τ::Int = 1) → Vector{<:Real}

Compute the unordered probabilities of the occurrence of symbol sequences constructed from the data x.

If x is a multivariate Dataset, then symbolization is performed directly on the state vectors. If x is a univariate signal, then a delay reconstruction with embedding lag τ and embedding dimension m is used to construct state vectors, on which symbolization is then performed.

A pre-allocated symbol array s can be provided to save some memory allocations if the probabilities are to be computed for multiple data sets. If provided, it is required that length(x) == length(s) if x is a Dataset, or length(s) == length(x) - (m-1)τ if x is a univariate signal.

See also: SymbolicPermutation.

Entropies.probabilities — Method

Weighted permutation-based symbol probabilities

probabilities(x::Dataset, est::SymbolicWeightedPermutation) → Vector{<:Real}  
probabilities(x::AbstractVector{<:Real}, est::SymbolicWeightedPermutation; m::Int = 3, τ::Int = 1) → Vector{<:Real}

probabilities!(s::Vector{Int}, x::Dataset, est::SymbolicWeightedPermutation) → Vector{<:Real}  
probabilities!(s::Vector{Int}, x::AbstractVector, est::SymbolicWeightedPermutation; m::Int = 3, τ::Int = 1) → Vector{<:Real}

Compute the unordered probabilities of the occurrence of weighted symbol sequences constructed from x.

If x is a multivariate Dataset, then symbolization is performed directly on the state vectors. If x is a univariate signal, then a delay reconstruction with embedding lag τ and embedding dimension m is used to construct state vectors, on which symbolization is then performed.

A pre-allocated symbol array s can be provided to save some memory allocations if the probabilities are to be computed for multiple data sets. If provided, it is required that length(x) == length(s) if x is a Dataset, or length(s) == length(x) - (m-1)τ if x is a univariate signal`.

See also: SymbolicWeightedPermutation.

Missing docstring.

Missing docstring for probabilities(x::Dataset, est::VisitationFrequency). Check Documenter's build log for details.